backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
email-validator==2.2.0
pypdf==4.3.1
rank_bm25==0.2.2
numpy==1.26.4
requests==2.32.3
reportlab==4.2.2
sentence-transformers==3.0.1
torch==2.4.1
faiss-cpu==1.8.0.post1
